[PC-876] refactor:DTO 위치 변경
