+++ admin/src/main/java/org/yapp/user/presentation/request/UpdateProfileStatusRequest.java
@@ -1,8 +1,8 @@
-package org.yapp.user.presentation;
+package org.yapp.user.presentation.request;
 
 import jakarta.validation.constraints.NotNull;
 
 public record UpdateProfileStatusRequest(@NotNull boolean rejectImage,
-                                         @NotNull boolean rejectDescription) {
+        @NotNull boolean rejectDescription) {
 
 }
